rename injected connectors
