--- packages/no-modal/src/connectors/default-evm-connector/index.ts
@@ -1,5 +1,0 @@
-export { createStore as createMipd } from "mipd";
-
-import { injectedEvmConnector } from "./injectedEvmConnector";
-
-export { injectedEvmConnector };
