+++ packages/no-modal/src/connectors/index.ts
@@ -2,6 +2,6 @@ export * from "./auth-connector";
 export * from "./base-evm-connector";
 export * from "./base-solana-connector";
 export * from "./coinbase-connector";
-export * from "./default-evm-connector";
-export * from "./default-solana-connector";
+export * from "./injected-evm-connector";
+export * from "./injected-solana-connector";
 export * from "./wallet-connect-v2-connector";
