+++ packages/no-modal/src/connectors/injected-evm-connector/index.ts
@@ -1,0 +1,2 @@
+export { injectedEvmConnector } from "./injectedEvmConnector";
+export { createStore as createMipd } from "mipd";
