+++ packages/no-modal/src/connectors/injected-solana-connector/index.ts
