+++ packages/no-modal/src/connectors/injected-solana-connector/utils.ts
